+++ vector/vector.go
@@ -201,7 +201,7 @@ func (z *Rasterizer) QuadTo(bx, by, cx, cy float32) {
 	devsq := devSquared(ax, ay, bx, by, cx, cy)
 	if devsq >= 0.333 {
 		const tol = 3
-		n := 1 + int(math.Sqrt(math.Sqrt(tol*float64(devsq))))
+		n := 1 + int(math.Sqrt(math.Sqrt(tol*float32(devsq))))
 		t, nInv := float32(0), 1/float32(n)
 		for i := 0; i < n-1; i++ {
 			t += nInv
@@ -225,7 +225,7 @@ func (z *Rasterizer) CubeTo(bx, by, cx, cy, dx, dy float32) {
 	}
 	if devsq >= 0.333 {
 		const tol = 3
-		n := 1 + int(math.Sqrt(math.Sqrt(tol*float64(devsq))))
+		n := 1 + int(math.Sqrt(math.Sqrt(tol*float32(devsq))))
 		t, nInv := float32(0), 1/float32(n)
 		for i := 0; i < n-1; i++ {
 			t += nInv
